build: move textbanner to recipes folder
