+++ packages/design-system/.storybook/main.js
@@ -1,6 +1,7 @@
 module.exports = {
   stories: [
     "../components/**/__stories__/*.stories.@(mdx|jsx)",
+    "../recipes/**/__stories__/*.stories.@(mdx|jsx)",
     "../icons/__stories__/*.stories.@(mdx|jsx)",
     "../theme/__stories__/*.stories.@(mdx|jsx)",
   ],
