+++ packages/design-system/recipes/text-banner/textBanner/TextBanner.styled.js
