+++ packages/design-system/recipes/text-banner/textBanner/__stories__/TextBanner.stories.mdx
@@ -2,13 +2,13 @@ import { Meta, Canvas, Story, ArgsTable } from "@storybook/addon-docs"
 
 import TextBanner from "../TextBanner.jsx"
 
-<Meta title="Components/TextBanner" component={TextBanner} />
+<Meta title="Recipes/TextBanner" component={TextBanner} />
 
 # Badge
 
 <Canvas>
   <Story
-    name="First"
+    name="Overview"
     args={{
       layout: "left",
       heading: "Pomóż nam pomagać",
