+++ packages/design-system/recipes/text-banner/textBanner/__tests__/TextBanner.test.jsx
@@ -1,4 +1,4 @@
-import { render, screen } from "../../../test-utils"
+import { render, screen } from "../../../../test-utils"
 
 import TextBanner from "../TextBanner"
